refactor(ui): extract SessionBrowser list components
